fix: ensure zoom slider uses double value
Agent-Logs-Url: https://github.com/science64/fast_barcode_scanner/sessions/19651f0f-42f6-4d3a-96be-e0019d629106

+++ fast_barcode_scanner/example/lib/scanner_screen.dart
@@ -93,7 +93,8 @@ class _ScannerScreenState extends State<ScannerScreen> {
                     final maxZoom = cameraState.maxZoom <= minZoom
                         ? minZoom + 0.01
                         : cameraState.maxZoom;
-                    final zoom = cameraState.zoom.clamp(minZoom, maxZoom);
+                    final zoom =
+                        cameraState.zoom.clamp(minZoom, maxZoom).toDouble();
 
                     return Column(
                       children: [
